backend -- books.* -- Добавление в модуль сборников ещё одну PDF - для АНГЛИЙСКОГО оглавления. Плюс добавлены конфирмы на удаление файлов из книг и удаление самих книг.

+++ www/core/core.filestorage.php
@@ -10,7 +10,10 @@ class FileStorage {
 
     public static function removeFile($id)
     {
-        mysql_query("DELETE FROM filestorage WHERE id=$id");
+        if ($id != -1)
+        {
+            mysql_query("DELETE FROM filestorage WHERE id=$id");
+        }
     }
 
     public static function getFile($id)
